+++ src/Math/Vector2.h
@@ -9,15 +9,15 @@ namespace Raycer
 	{
 	public:
 
-		Vector2(float x = 0.0f, float y = 0.0f);
+		Vector2(double x = 0.0, double y = 0.0);
 		Vector2(const Vector2& v);
 		Vector2& operator=(const Vector2& v);
 
 		friend Vector2 operator+(const Vector2& v, const Vector2& w);
 		friend Vector2 operator-(const Vector2& v, const Vector2& w);
-		friend Vector2 operator*(const Vector2& v, float s);
-		friend Vector2 operator*(float s, const Vector2& v);
-		friend Vector2 operator/(const Vector2& v, float s);
+		friend Vector2 operator*(const Vector2& v, double s);
+		friend Vector2 operator*(double s, const Vector2& v);
+		friend Vector2 operator/(const Vector2& v, double s);
 		friend Vector2 operator-(const Vector2& v);
 
 		friend bool operator==(const Vector2& v, const Vector2& w);
@@ -25,20 +25,20 @@ namespace Raycer
 
 		Vector2& operator+=(const Vector2& v);
 		Vector2& operator-=(const Vector2& v);
-		Vector2& operator*=(float s);
-		Vector2& operator/=(float s);
+		Vector2& operator*=(double s);
+		Vector2& operator/=(double s);
 
-		float length() const;
-		float lengthSquared() const;
+		double length() const;
+		double lengthSquared() const;
 		void normalize();
 		Vector2 normalized() const;
 		bool isZero() const;
 		bool isNormal() const;
-		float dot(const Vector2& v) const;
+		double dot(const Vector2& v) const;
 		Vector2 reflect(const Vector2& normal) const;
-		static Vector2 lerp(const Vector2& v1, const Vector2& v2, float t);
+		static Vector2 lerp(const Vector2& v1, const Vector2& v2, double t);
 
-		float x;
-		float y;
+		double x;
+		double y;
 	};
 }
